Fixed issue with no version cli param
There was an issue with the script where it had a bad link when no
version option was passed to the script.  The issue was due to
formatting of if-statements which checked if a version had been
specified.  Now, the script will install latest version of the jlink
software package if no version option is passed to script.