Working Docker Image
Docker image successfully installs JLink software package;  creates
generic docker image based on ubunutu.  The user is called dev, and the
container starts in dev's home directory.  container_entrypoint.sh adds
the JLink binaries to the $PATH and runs whatever CMD is passed
to docker (for example when CMD='/bin/bash', the command "docker run
-it anvil '/bin/bash'" executes as expected).  A '--name-only' flag was
added to the get_jlink_... script so that the dockerfile can know which
file to extract.