Moved USB config flags to compose file
Moved the bind mount flag and the device_cgroup_rules flag to the
docker compose file.  This means new containers have sensible default
configs.

Running the containers using docker-compose introduces an issue:  the
command must be run in the same directory as the 'compose.yaml', or the
command must be passed the path to 'compose.yaml'.  Can fix in future by
encapsulating commands with python (or something similliar) wrapper.